Create TRACK_BITMAP in map_create
This way, we don't have to guess the dimensions etc., but we use the
same parameters that are used for MAP creation. Somehow, the
parameters we used previously were slightly off, when used with
the Ruudskogen map. The track bitmap had "lines" over the track due to
smaller scale. Now, the Ruudskogen track in contiguous.